monitoring/requirements.txt
dataclasses==0.6
Flask~=2.0.1
pandas~=1.1.5
Werkzeug~=2.0.1
requests~=2.26.0
prometheus_client~=0.11.0
pyyaml~=5.4.1
pyarrow
pymongo
